requirements.txt
pillow
pytube
moviepy
requests